requirements.txt
celery==5.0.0
coloredlogs==14.0
coverage==5.3
cryptography==3.3.2
dj-database-url==0.5.0
django-admin-thumbnails==0.2.5
django-celery-beat==2.0.0
django-celerybeat-status==0.0.10
django-cors-headers==3.5.0
django-filter==2.3.0
django-rest-framework==0.1.0
django-timezone-field==4.0
django-webpack-loader==0.7.0
Django==3.1.13
djangorestframework-queryfields==1.0.0
djangorestframework-simplejwt==4.4.0
djangorestframework==3.11.2
drf-url-filters==0.5.1
editdistance==0.3.1
ExifRead==2.3.1
flower==0.9.7
gpsphoto==2.2.3
gunicorn==20.0.4
imageio==2.9.0
importlib-metadata==2.0.0
matplotlib==3.3.2
numpy==1.20.1
pgi==0.0.11.2
piexif==1.1.3
Pillow==8.3.2
postgres==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-xmp-toolkit==2.0.1
pytz==2020.1
redis==3.5.3
regex==2020.7.14
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
SCons==4.0.1
torch==1.8.0
torchvision==0.9.0
twilio==6.45.3
xmltodict==0.12.0
